13: slide in on scroll
1. When adding an event listener on scroll, it logs so many entries per
second. Therefore, we `debounce` it, to pay attention to scrolling every so
seconds. We pass it a function and `debounce` will run all the time,
but will only call the passed function (in our case `checkSlide)
every 20 miliseconds (`wait` time).
2. In `checkSlide` function, we loop over every image,
and will need to decide when it needs to be shown,
when it starts picking out at 50% of its height.
3. `window.scrollY` gives up the position of the top of the browser on
the web page. `window.scrollY + window.innerHeight` gives us the bottom
of the browser window.
4. For `slideInAt` we grab the position of the bottom of the page,
and subtract half of the height of the image => to know the point
(pixel value on the page) when the image should start sliding in.
5. `slideImage.offsetTop` => how far is the top of the image from the
top of the window. But to know where the bottom of the image is, we need
to add the `slideImage.height`.