14: JS references vs copying
1. Reassigning variables means you need to use `let`
2. Changing the original value of strings, numbers, and booleans works
for the variable.
3. in Line 27, we update `team`, which references `players`. Thus
changing `team`, also changes the original `players`. When updating an
Array, it will always reference back.
4. We can create a copy of the Array by using, `slice()`,
`[].concat(oldArray)`, ES6 spread, or `Array.from(players)`
5. Spread will take every items in the old Array and put a copy of them
in the new array.
6. Same thing goes for objects as well. But to copy an object, we start
with an `Object.assign()`, which takes an empty object and the
properties of the object to be copied. And lastly, you pass (fold) in
the properties you wish to change: `Object.assign({}, person, { age: 99
});` (line 64)
7. `Object.assign()` copies one level deep. Any nested levels, when you
update the new object, will also update the original object. Instead
need to look into deep cloning to clone every level of the object.
8. A cheat way to solve this problem is to use:
`JSON.parse(JSON.stringify(obj))`. So we convert the object to a string,
and immediately, parse it to an object using the `JSON.parse`, which
gives you a full copy. But should look into performance before using it.