15: persisting state with local storage part 1
- We add: `e.preventDefault()`, because on form submit, the page gets
  refreshed and we don't want that.
- We can grab the value of the item's name from the text field, by query
  the selector within the form element that was passed to the function.
It narrows down the search for the element that we are looking for to
the form's elements. Then we wrap it in parans, to be able to get its
value.
- When defining the item, we can now pass the value of the `text` and
  assign it to `text` (line 32 i.e. `text: text,`) but we can also use
the IE6 shorthand and remove the explicit assignment, and it will still
work.
- Once the form is submitted, we want to reset it, so we don't keep
  submitting the same information again and again (line 38).
- Before reseting the form, we want to push the item into the `items`
  array (line 36).
- In `populateList` function, we pass the array of items, which here we
  call `plates` and default it to an empty array, so if nothing was
passed, the function will still loop through an empty array, but at
least it will not throw an error. We also pass an HTML element to
populate the items in, which we call `platesList`.
- We use `map().join("")` because map() returns an array, but we need a
  long string to be able to rewrite the inner html of `platesList`.
- In this case, every time we call `populateList` it recreates the whole
  list. For better performance, we would like to update just that one
last item, something similar to jquery's `append` (would check what's
the minimum html that needs to be change on the page).
- Because we store all the data in an Array called `items`, when we
  refresh the page, all the items are gone. So we need to use local
storage to persist them, which will be in the next commit...