19: web cam
1. package.json has one dev dependency: `browser-sync`, which allows us
to open up a web server and live reloading. We can install it with `npm
install`, followed by `npm start`. It will kick open a local server. It
also gives us details for an external server so that we can check how
the web app works on the wifi.

2. Video from the web cam will come in, and will be piped in into the
video element. Snapshots from that video will be dumped into the canvas
element. Once in canvas, we can manipulate the photos by applying
filters, and more.

3. `navigator.mediaDevices.getUserMedia({ video: true, audio: false })`
(line 8) returns a promise and we use `.then()` on promises.

4. We cannot assign the `localMediaStream` to `video.src` because for it
to work it needs to be converted to a url when dealing with a live
stream. This explains using the following on line 11:
`window.URL.createObjectURL(localMediaStream)`. It will convert the live
stream to something video can understand.

5. When we inspect the video source in Elements tab, we can see that the
video element is a blob, the raw data is being piped in.

6. In case we don't allow the browser to use the web cam, we need a
`.catch()` clause (line 14)