22: follow along links
- When `mouseenter` on a link, we can call
  `this.getBoundingClientRect()` to get (as the name suggests) the
bounds of the client. So we get a `ClientRect` object, which has the
following attributes: bottom, height, left, right, top, and width.
Basically, where on the page this element lives.

- Then we can grab these attributes and use them to style the
  `span.highlight` (line 46-47), which has a transition defined on it
(styles.css:24). Without the transition, the size of the highlight span
would change in immediate snaps, which is not what we want. We want it
to animate itself.

- But we have a problem, when scrolling down, the highlighted span's
  position is off. So we can figure out what the offset is. But a second
solution is to figure out how far down we scrolled down and take it off
the element's position.

- `window.scrollY` gives us how far down we scrolled on the page. So we
  use that to create our own custom coordinates (line 39-44).