24: sticky nav
- When scrolling the page down, when the top of the nav touches the top
  of the browser window, `nav.offsetTop` and `window.scrollY` will be
the same. This is the point that we want to make the nav sticky.

- When we make the nav `position: fixed`, it no longer takes space on
  the page (it sort of floats on top of the page), and it causes a
reflow on the page so that the content jumps up to make up the space
that just freed up. We can counteract that, by adding some padding to
our body, just as we make the nav fixed. (line 244). The padding height
is taken from the nav element.